pyproject.toml
[project]
name = "hello-world"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.84.0",
    "openai-agents[litellm]>=0.0.17",
    "python-dotenv>=1.1.0",
]
